Add fastsd API client
